CXFunction.go: Collect definition of and methods on CXFunction
