+++ cx/removers.go
@@ -18,52 +18,6 @@ func (strct *CXStruct) RemoveField(fldName string) {
 	}
 }
 
-func (fn *CXFunction) RemoveExpression(line int) {
-	if len(fn.Expressions) > 0 {
-		lenExprs := len(fn.Expressions)
-		if line >= lenExprs-1 || line < 0 {
-			fn.Expressions = fn.Expressions[:len(fn.Expressions)-1]
-		} else {
-			fn.Expressions = append(fn.Expressions[:line], fn.Expressions[line+1:]...)
-		}
-		// for i, expr := range fn.Expressions {
-		// 	expr.Index = i
-		// }
-	}
-}
-
-func (fn *CXFunction) RemoveInput(inpName string) {
-	if len(fn.Inputs) > 0 {
-		lenInps := len(fn.Inputs)
-		for i, inp := range fn.Inputs {
-			if inp.Name == inpName {
-				if i == lenInps {
-					fn.Inputs = fn.Inputs[:len(fn.Inputs)-1]
-				} else {
-					fn.Inputs = append(fn.Inputs[:i], fn.Inputs[i+1:]...)
-				}
-				break
-			}
-		}
-	}
-}
-
-func (fn *CXFunction) RemoveOutput(outName string) {
-	if len(fn.Outputs) > 0 {
-		lenOuts := len(fn.Outputs)
-		for i, out := range fn.Outputs {
-			if out.Name == outName {
-				if i == lenOuts {
-					fn.Outputs = fn.Outputs[:len(fn.Outputs)-1]
-				} else {
-					fn.Outputs = append(fn.Outputs[:i], fn.Outputs[i+1:]...)
-				}
-				break
-			}
-		}
-	}
-}
-
 func (expr *CXExpression) RemoveInput() {
 	if len(expr.Inputs) > 0 {
 		expr.Inputs = expr.Inputs[:len(expr.Inputs)-1]
